Update dockerfile patch and make new java devfiles used patched images
Update the arbitrary users patch to make sure $HOME is always writable,
fixing issues with some images.

Update java-spring, java-mysql, and java-vertx to use patched images
  - Update the commands in these images since maven wrapper is
    incompatible with community maven image

+++ arbitrary-users-patch/entrypoint.sh
@@ -6,7 +6,7 @@ if [ ! -d "${HOME}" ]; then
 fi
 
 # Setup $PS1 for a consistent and reasonable prompt
-if [ ! -f "${HOME}"/.bashrc ]; then
+if [ -w "${HOME}" ] && [ ! -f "${HOME}"/.bashrc ]; then
   echo "PS1='\s-\v \w \$ '" > "${HOME}"/.bashrc
 fi
